wrap segment logic into a utility function and optimize imports in mdast-to-react and wrap-segments components
