+++ next/src/app/[locale]/_components/mdast-to-react.tsx
@@ -1,23 +1,15 @@
-import { SegmentWrapper } from "@/app/[locale]/_components/SegmentWrapper";
 import type { SegmentBundle } from "@/app/[locale]/types";
 import type { Prisma } from "@prisma/client";
 import Image from "next/image";
-import {
-	type ComponentType,
-	type JSX,
-	type ReactElement,
-	createElement,
-} from "react";
-import React from "react";
-import { Tweet as XPost } from "react-tweet";
+import type { ReactElement } from "react";
+import { type ComponentType, type JSX, createElement } from "react";
 import * as jsxRuntime from "react/jsx-runtime";
 import rehypeRaw from "rehype-raw";
 import rehypeReact from "rehype-react";
 import rehypeSlug from "rehype-slug";
 import remarkRehype from "remark-rehype";
 import { unified } from "unified";
-
-const TWEET_ID_RE = /status\/(\d+)/;
+import { wrapSegment } from "./wrap-segments";
 
 const SEGMENTABLE = [
 	"p",
@@ -55,55 +47,6 @@ const Link: ComponentType<JSX.IntrinsicElements["a"]> = ({
 		{children}
 	</a>
 );
-
-function wrapSegment<Tag extends keyof JSX.IntrinsicElements>(
-	Tag: Tag,
-	bundles: SegmentBundle[],
-	current?: string,
-) {
-	return (p: JSX.IntrinsicElements[Tag] & { "data-number-id"?: number }) => {
-		/* ───────── ここから追加ロジック ───────── */
-		// <p> の唯一の子が tweet リンクなら <TweetContainer> で置換
-		if (Tag === "p") {
-			const kids = React.Children.toArray(p.children);
-			if (
-				kids.length === 1 &&
-				React.isValidElement<{ href?: string }>(kids[0]) &&
-				TWEET_ID_RE.test(kids[0].props.href ?? "")
-			) {
-				const id = TWEET_ID_RE.exec(kids[0].props.href ?? "")?.[1];
-				if (id) return <XPost id={id} />;
-			}
-		}
-		const id = p["data-number-id"];
-		const bundle =
-			id !== undefined
-				? bundles.find((b) => b.segment.number === +id)
-				: undefined;
-
-		/* ───────── 追加ここまで ───────── */
-		/* セグメント対象でなければそのまま DOM 要素を返す */
-		if (!bundle) return createElement(Tag, p, p.children);
-
-		/* --- ここで Client Component に “シリアライズ可能な形” で渡す --- */
-		const { children, ...rest } = p;
-		return (
-			<SegmentWrapper
-				bundle={bundle}
-				tagName={Tag}
-				tagProps={rest as JSX.IntrinsicElements[Tag]} // ★そのまま突っ込む
-				currentHandle={current}
-			>
-				{children} {/* ← children は React が面倒見てくれる */}
-			</SegmentWrapper>
-		);
-	};
-}
-
-/* -------------------------------------------------------------------------- */
-/*                               Main Converter                               */
-/* -------------------------------------------------------------------------- */
-
 interface Params {
 	mdast: Prisma.JsonValue;
 	bundles: SegmentBundle[];
