+++ next/src/app/[locale]/_components/wrap-segments.tsx
@@ -1,65 +1,51 @@
 "use client";
+import { SegmentWrapper } from "@/app/[locale]/_components/SegmentWrapper";
 import type { SegmentBundle } from "@/app/[locale]/types";
-import { useDisplay } from "@/app/_context/display-provider";
-import { Fragment, createElement } from "react";
-import type { JSX } from "react/jsx-runtime";
-import { TranslationSection } from "./segment-and-translation-section/translation-section";
+import { type JSX, createElement } from "react";
+import React from "react";
+import { Tweet as XPost } from "react-tweet";
+const TWEET_ID_RE = /status\/(\d+)/;
 
 export function wrapSegment<Tag extends keyof JSX.IntrinsicElements>(
 	Tag: Tag,
 	bundles: SegmentBundle[],
 	current?: string,
 ) {
-	return (
-		allProps: JSX.IntrinsicElements[Tag] & { "data-number-id"?: number },
-	) => {
-		const { children, ...rest } = allProps;
-		const id = allProps["data-number-id"];
+	return (p: JSX.IntrinsicElements[Tag] & { "data-number-id"?: number }) => {
+		/* ───────── ここから追加ロジック ───────── */
+		// <p> の唯一の子が tweet リンクなら <TweetContainer> で置換
+		if (Tag === "p") {
+			const kids = React.Children.toArray(p.children);
+			if (
+				kids.length === 1 &&
+				React.isValidElement<{ href?: string }>(kids[0]) &&
+				TWEET_ID_RE.test(kids[0].props.href ?? "")
+			) {
+				const id = TWEET_ID_RE.exec(kids[0].props.href ?? "")?.[1];
+				if (id) return <XPost id={id} />;
+			}
+		}
+		const id = p["data-number-id"];
 		const bundle =
 			id !== undefined
 				? bundles.find((b) => b.segment.number === +id)
 				: undefined;
 
-		/* ── セグメント対象でなければそのまま ── */
-		if (!bundle) return createElement(Tag, allProps, children);
+		/* ───────── 追加ここまで ───────── */
+		/* セグメント対象でなければそのまま DOM 要素を返す */
+		if (!bundle) return createElement(Tag, p, p.children);
 
-		/* ── 表示モード＆色 ───────────────────── */
-		const { mode } = useDisplay();
-		const hasTr = bundle.translations.length > 0;
-		const effMode = mode === "user" && !hasTr ? "source" : mode;
-
-		const gray = "text-gray-300 dark:text-gray-600";
-		const normal = "text-gray-700 dark:text-gray-200";
-		const srcColor = effMode !== "source" && hasTr ? gray : normal;
-
-		/* ── オリジナル (=source) ── */
-		const src =
-			effMode !== "user"
-				? createElement(
-						Tag,
-						{
-							...rest,
-							className: [rest.className, srcColor].filter(Boolean).join(" "),
-						},
-						children,
-					)
-				: null;
-
-		/* ── 翻訳 (=TranslationSection) ── */
-		const tr =
-			effMode !== "source" && hasTr
-				? createElement(
-						Tag,
-						{ ...rest, key: `tr-${id}` },
-						<TranslationSection
-							segmentBundle={bundle}
-							currentHandle={current}
-							interactive
-						/>,
-					)
-				: null;
-
-		/* ── 同じタグで上下に並べる ── */
-		return createElement(Fragment, null, src, tr);
+		/* --- ここで Client Component に “シリアライズ可能な形” で渡す --- */
+		const { children, ...rest } = p;
+		return (
+			<SegmentWrapper
+				bundle={bundle}
+				tagName={Tag}
+				tagProps={rest as JSX.IntrinsicElements[Tag]} // ★そのまま突っ込む
+				currentHandle={current}
+			>
+				{children} {/* ← children は React が面倒見てくれる */}
+			</SegmentWrapper>
+		);
 	};
 }
